requirements.txt
streamlit
transformers
sentence-transformers
pinecone-client
pinecone-text
langchain
langchain_community
openai
protobuf
datasets